Introduce flag that allows listening on an arbitrary interface.
Care is taken that the old behavior which uses the `--port` flag is
preserved. If the `--port` flag does not have the default, empty
value, then its value is used for the new `--httpaddr` flag, prefixed
with `127.0.0.1:`.

(Strictly speaking, if `localhost` does not resolve to `127.0.0.1`,
then the behavior is not identical.)

Fixes gogo/letmegrpc#15.